pyproject.toml
[project]
name = "a7-lang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.2",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.4.8",
    "langsmith>=0.4.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]
